+++ python/mujoco/specs.cc
@@ -1568,6 +1568,23 @@ PYBIND11_MODULE(_specs, m, pybind11::mod_gil_not_used()) {
       py::arg("kp"), py::arg("kv") = -1, py::arg("dampratio") = -1,
       py::arg("ctrlspec") = 0);
   mjsActuator.def(
+      "set_to_pid",
+      [](raw::MjsActuator* self, double kp, double kv, double dampratio,
+         double ki, double imax, double slewmax, double inheritrange,
+         int ctrlspec) {
+        std::string err = mjs_setToPID(
+            self, kp, kv == -1 ? nullptr : &kv,
+            dampratio == -1 ? nullptr : &dampratio, ki == -1 ? nullptr : &ki,
+            imax == -1 ? nullptr : &imax, slewmax == -1 ? nullptr : &slewmax,
+            inheritrange, ctrlspec);
+        if (!err.empty()) {
+          throw pybind11::value_error(err);
+        }
+      },
+      py::arg("kp"), py::arg("kv") = -1, py::arg("dampratio") = -1,
+      py::arg("ki") = -1, py::arg("imax") = -1, py::arg("slewmax") = -1,
+      py::arg("inheritrange") = 0, py::arg("ctrlspec") = 0);
+  mjsActuator.def(
       "set_to_damper",
       [](raw::MjsActuator* self, double kv) {
         std::string err = mjs_setToDamper(self, kv);